Multiple cycle types and graphIV directory selection
Select a directory in graphIV and it will process everything inside. Right now it just processes all CSVs. Need to find a way to get it to only try and process real log files.
Multiple cycle types can be added to a single test from the GUI, and added a rest type.